Moved to list

+++ internal/model/training/model.go
@@ -15,7 +15,7 @@ type Training struct {
 	StartedAt      time.Time
 	EndedAt        time.Time
 	AdditionalInfo string
-	Duration       int
+	Duration       int64
 }
 
 type CreateTrainingEvent struct {
